fix(take-sizing): exact integer arithmetic, so the money path is ABI-independent
CI's Ubuntu job failed three assertions in test_take_sizing.cpp that pass on
MSVC. long double is 64-bit on MSVC and 80-bit x87 on GCC/x86-64, and both are
too narrow here: the true answers are exactly 1500, 750 and 2000000000000, GCC
returns 1501, 751 and 2000000000001, and MSVC is right only by luck of a
different rounding. The tests were correct and the code was wrong, so no
assertion was loosened.

Replaced with exact integer arithmetic: ceil(n1*n2*n3 / (d1*d2)), saturating,
in pure ISO C++20 -- no float, no __int128, no intrinsic, no #if, one code path
on every platform. 64x64->128 via 32-bit limbs, 128x64->192, then restoring
binary long division. Two separate 64-bit divisions rather than one wide
divisor, which is forced: on XCH/DBX the combined denominator is 1e24, itself
80 bits. Exactness is recovered because N mod (d1*d2) == 0 iff both remainders
are zero.

128 bits is not enough, which rules out __int128 on correctness grounds and not
merely portability: quote_size * kMojosPerXch * base_mpu reaches ~2^166, and an
unremarkable bid of 1e15 quote mojos on an XCH-base pair already needs 130.
Pinned by a test.

The whole thing is constexpr and the header carries fourteen static_asserts.
That is the answer to "GCC cannot be run locally": GCC evaluates them at build
time, so a wrong answer becomes a compile error rather than a red runner, and
constant evaluation forbids UB, so a bad shift fails everywhere.

The anti-regression test is the part worth keeping. quote_cost_for_ask(
10'000'001, 99'999'990'000'001, 1e12, 1e3) has base*price = 10^21 + 1 exactly,
70 significand bits, so the true ceiling is 2 -- and BOTH 53-bit double and
64-bit x87 return 1. Verified independently by emulating both formats. The
mutation "revert to long double" is therefore caught on MSVC, the ABI where the
original defect was invisible.

SCOPE EXTENDED, deliberately and flagged: engine.cpp carried two more verbatim
copies of the same defective formula (Step 8, the auto-rebalance proactive
cancel and exposure_breaches_reserve -> can_bid/can_ask, i.e. a quoting kill
switch). Both now call the shared function. Fixing only take_sizing.hpp would
have left two live copies of a bug we had just declared fixed.

Behavioural delta is +/-1 mojo on divergent inputs. Note the direction: the old
MSVC code sometimes returned one mojo BELOW the true ceiling, which is the
fail-open direction inside a guard whose only job is to prevent fail-open. On
the live XCH/DBX pair I could find no realistic input where double diverges
(0 of 17,500 sampled); wmilliETH.b/XCH diverges in 3.6% of cases and is
disabled. So nothing has been mis-spent -- the argument for this change is that
the divergence set depends on the ABI and neither contains the other, so no
reasoning about which inputs are safe was possible at all.

Also corrected: my own diagnosis said the loss was at base*price needing 81
bits. It is 58 bits, which x87 holds fine; the loss is one operation later, at
* quote_mpu, which needs 65 against x87's 64. Conclusion unchanged, reasoning
was wrong.

1,161 tests pass, zero warnings at /W4 /WX, all static_asserts satisfied at
compile time. Nine mutations applied and all nine killed. xop_trader never
built -- still 2026-08-30 13:27:45, 2,696,704 bytes.

RESIDUAL RISK: GCC warnings only. The values are proven identical by
construction, but -Wall -Wextra -Wpedantic -Werror has not been observed on
this code. That is what the CI run is for.

+++ cpp/include/xop/execution/crossed_book.hpp
@@ -140,11 +140,12 @@ namespace xop::execution {
     if (base_mojos_per_unit <= 0) return 0;
     if (!(max_take_units > 0.0) || !std::isfinite(max_take_units)) return 0;
 
-    // 9.0e18 is exactly representable as a double (and long double is double
-    // under MSVC), and is strictly below 2^63, so both the comparison and the
-    // subsequent narrowing are exact and well-defined on every ABI we build.
-    // A previous saturation clamp in this repo rounded to 2^63 and passed
-    // MSVC while failing GCC; do not replace this with (long double)INT64_MAX.
+    // 9.0e18 is exactly representable as a double (9e18 = 2^18 * 34332275390625,
+    // a 46-bit odd part) and is strictly below 2^63, so both the comparison
+    // and the subsequent narrowing are exact and well-defined on every ABI we
+    // build. A previous saturation clamp in this repo rounded to 2^63 and
+    // passed MSVC while failing GCC; do not replace this with
+    // (long double)INT64_MAX.
     //
     // NO TEST GUARDS THIS LINE, and that was checked rather than assumed:
     // deleting it leaves the whole suite green on MSVC. Narrowing an
@@ -154,10 +155,44 @@ namespace xop::execution {
     // it is ABI-specific. THIS BOUND is what makes the conversion defined;
     // the `cap > 0` below is a second line of defence, not the first. Do not
     // delete it on the grounds that the tests still pass.
-    constexpr long double kMaxSafe = 9.0e18L;
+    //
+    // [2026-09-01] THE TYPE HERE IS `double`, DELIBERATELY, AND IT USED TO BE
+    // `long double`. The comment above was true of the NARROWING and false of
+    // the VALUE, which is a distinction worth spelling out because this
+    // function is cited elsewhere as the exemplar of getting this right.
+    //
+    // `long double` is 53-bit on MSVC and 64-bit x87 on GCC/x86-64, so
+    // `units * base_mojos_per_unit` was computed to different precisions and
+    // then TRUNCATED -- the cap VALUE was ABI-divergent, in the same file
+    // that warns about ABI divergence. For max_take_units = 0.15 and
+    // base_mojos_per_unit = 1e12 the exact product is 149999999999.999994...,
+    // which the 53-bit rounding lifts to exactly 150000000000 while the
+    // 64-bit rounding leaves below it: MSVC 150000000000, GCC 149999999999.
+    // Same for 0.03 and 0.3. This is the cap that `base_sz > max_mojos` is
+    // tested against at Step 9e and Step 9f -- the same guard the exact
+    // base_size_for_bid result feeds.
+    //
+    // It was latent, not live: every current config value
+    // (crossed_book_max_take_xch 5.0, midpoint_recycling_max_take_xch 0.25,
+    // peg_arb_max_take_units 50, drift_corrector_max_take_units 2.0) is
+    // dyadic-exact and agrees on both ABIs. Any operator writing 0.15, 0.03
+    // or 0.3 would have made CI and production disagree by one mojo again.
+    //
+    // `double` fixes it because binary64 arithmetic is correctly rounded and
+    // identical on both toolchains here: GCC targets x86-64 with SSE2, where
+    // FLT_EVAL_METHOD is 0, so there is no x87 excess precision to leak in.
+    // MSVC is bit-identical to before (its long double IS double), so the
+    // DEPLOYED build's cap values do not move at all.
+    //
+    // Note this is portable rounding, NOT exact-rational evaluation, and that
+    // is the intended reading: 0.15 * 1e12 in exact rational over the double
+    // 0.15 is 149999999999, whereas round-to-nearest recovers the 150000000000
+    // the operator meant when they typed 0.15. Truncation afterwards stays --
+    // it only ever narrows the cap, which is the conservative direction.
+    constexpr double kMaxSafe = 9.0e18;
 
-    const long double cap_ld = static_cast<long double>(max_take_units)
-                             * static_cast<long double>(base_mojos_per_unit);
+    const double cap_ld = static_cast<double>(max_take_units)
+                        * static_cast<double>(base_mojos_per_unit);
     if (cap_ld > kMaxSafe) return 0;
 
     const Mojo cap = static_cast<Mojo>(cap_ld);  // truncates down -- conservative
